% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{al122005_prblty}
\alias{al122005_prblty}
\title{Strike probabilities for Hurricane Katrina (AL122005)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 937 rows and 10 columns.}
\source{
\url{http://www.nhc.noaa.gov/archive/2005/KATRINA.shtml?}
}
\usage{
al122005_prblty
}
\description{
Strike probabilities for Hurricane Katrina (AL122005)
}
\keyword{datasets}
