% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_chart.R
\name{al_tracking_chart}
\alias{al_tracking_chart}
\title{al_tracking_chart}
\usage{
al_tracking_chart(...)
}
\value{
ggplot2 object centered on Atlantic basin.
}
\description{
Build tracking chart centered on Atlantic Basin.
}
\seealso{
\code{\link{tracking_chart}}
}
