% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storms.R
\name{build_archive_df}
\alias{build_archive_df}
\title{build_archive_df}
\usage{
build_archive_df(link, basin = c("AL", "EP"))
}
\arguments{
\item{basin}{c("AL", "EP")}

\item{page}{contents of page being parsed}
}
\value{
dataframe
}
\description{
Add storms for the given link, basin into dataframej.
}
\keyword{internal}
