% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_chart.R
\name{build_tracking_chart_df}
\alias{build_tracking_chart_df}
\title{build_tracking_chart_df}
\usage{
build_tracking_chart_df(dataset, package)
}
\arguments{
\item{dataset}{Name of dataset to retrieve}

\item{package}{Package to which the requested dataset belongs.}
}
\description{
Helper function to retrieve appropriate dataset from the
    rnaturalearth data packages.
}
\keyword{internal}
