% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_lat_lon}
\alias{convert_lat_lon}
\title{convert_lat_lon}
\usage{
convert_lat_lon(x, y)
}
\arguments{
\item{x}{integer}

\item{y}{character}
}
\value{
integer positive or negative
}
\description{
Converts lat, lon to negative if in southern, western hemisphere, respectively
}

