% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{create_df_archives}
\alias{create_df_archives}
\title{create_df_archives}
\usage{
create_df_archives()
}
\value{
dataframe
}
\description{
Template for archives dataframe
\describe{
  \item{Year}{integer, year of storm}
  \item{Name}{character, name of storm}
  \item{Basin}{character, basin storm developed in}
  \item{Link}{character, link to storm's archive page}
}
}
\keyword{internal}
