% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discus.R
\name{create_df_discus}
\alias{create_df_discus}
\title{create_df_discus}
\usage{
create_df_discus()
}
\value{
empty dataframe
}
\description{
Template for storm discussions dataframe
}
\seealso{
\code{\link{get_discus}}
}
\keyword{internal}
