% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posest.R
\name{create_df_posest}
\alias{create_df_posest}
\title{create_df_posest}
\usage{
create_df_posest()
}
\value{
empty dataframe
}
\description{
Template for position estimates dataframe
}
\seealso{
\code{\link{get_posest}}
}
\keyword{internal}
