% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{create_df_prblty}
\alias{create_df_prblty}
\title{create_df_prblty}
\usage{
create_df_prblty()
}
\value{
empty dataframe
}
\description{
Template for strike probabilities dataframe
}
\seealso{
\code{\link{get_prblty}}
}
\keyword{internal}
