% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{create_df_public}
\alias{create_df_public}
\title{create_df_public}
\usage{
create_df_public()
}
\value{
empty dataframe
}
\description{
Template for public advisory dataframe
}
\seealso{
\code{\link{get_public}}
}
\keyword{internal}
