% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{create_df_update}
\alias{create_df_update}
\title{create_df_update}
\usage{
create_df_update()
}
\value{
empty dataframe
}
\description{
Template for cyclone update dataframe
}
\seealso{
\code{\link{get_update}}
}
\keyword{internal}
