% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{create_df_updates}
\alias{create_df_updates}
\title{create_df_updates}
\usage{
create_df_updates()
}
\value{
empty dataframe
}
\description{
Template for cyclone updates dataframe
}
\seealso{
\code{\link{get_updates}}
}
\keyword{internal}
