% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wndprb.R
\name{create_df_wndprb}
\alias{create_df_wndprb}
\title{create_df_wndprb}
\usage{
create_df_wndprb()
}
\value{
empty dataframe
}
\description{
Template for wind probabilities dataframe
}
\seealso{
\code{\link{get_wndprb}}
}
\keyword{internal}
