% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crul.R
\name{crul_discus}
\alias{crul_discus}
\title{crul_discus}
\usage{
crul_discus(contents)
}
\arguments{
\item{contents}{text of Storm Discussion product}
}
\description{
Parse storm Discussion products
}
\keyword{internal}
