% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crul.R
\name{crul_extract_storms}
\alias{crul_extract_storms}
\title{crul_extract_storms}
\usage{
crul_extract_storms(basin, contents)
}
\arguments{
\item{basin}{AL or EP}

\item{contents}{Contents of archive pages}
}
\description{
Extract storms for the given basin
}
