% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crul.R
\name{crul_posest}
\alias{crul_posest}
\title{crul_posest}
\usage{
crul_posest(contents)
}
\arguments{
\item{contents}{URL of a specific position estimate product}
}
\description{
Extrapolate data from Position Estimate products.
}
