% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crul.R
\name{crul_wndprb}
\alias{crul_wndprb}
\title{crul_wndprb}
\usage{
crul_wndprb(contents)
}
\arguments{
\item{contents}{Link to a storm's specific wind probability product.}
}
\description{
Parse wind probability products
}
