% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.al_2012}
\alias{df.al_2012}
\title{Atlantic cyclones for 2012}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 4 columns.
}
\source{
\url{http://www.nhc.noaa.gov/archive/2012/}
}
\usage{
df.al_2012
}
\description{
Atlantic cyclones for 2012
}
\keyword{datasets}
