% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.gis_adv}
\alias{df.gis_adv}
\title{GIS advisory dataset for Hurricane Sandy Adv 18}
\format{
An object of class \code{list} of length 4.
}
\source{
\url{http://www.nhc.noaa.gov/gis/archive_forecast_results.php?id=al18&year=2012&name=Hurricane\%20SANDY}
}
\usage{
df.gis_adv
}
\description{
GIS advisory dataset for Hurricane Sandy Adv 18
}
\keyword{datasets}
