% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discus.R
\name{discus}
\alias{discus}
\title{discus}
\usage{
discus(link, display_link = TRUE)
}
\arguments{
\item{link}{Link to a storm's specific discussion product.}

\item{display_link}{Display each link as being worked; default is TRUE.}
}
\value{
Dataframe
}
\description{
Parse storm Discussion products
}
\details{
Given a direct link to a discussion product, parse and return 
dataframe of values.
}
\seealso{
\code{\link{get_discus}}
}

