% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wndprb.R
\name{ep_prblty_stations}
\alias{ep_prblty_stations}
\title{ep_prblty_stations}
\usage{
ep_prblty_stations()
}
\description{
Retrieve list of probability stations based in the eastern
Pacific from the NHC. To be used in tandem with `wndprb` products.
}
\details{
This is a placeholder function. The current listing does not match
the format for Atlantic and central Pacific stations.

This function may be deprecated soon as the data sources may be
removed from the National Hurricane Center.
}
