% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_chart.R
\name{ep_tracking_chart}
\alias{ep_tracking_chart}
\title{ep_tracking_chart}
\usage{
ep_tracking_chart(...)
}
\value{
ggplot2 object centered on northeast Pacific basin.
}
\description{
Build tracking chart centered on northeast Pacific Basin.
}
\seealso{
\code{\link{tracking_chart}}
}
