% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storm_data.R
\name{extract_product_contents}
\alias{extract_product_contents}
\title{extract_product_contents}
\usage{
extract_product_contents(links, products)
}
\arguments{
\item{links}{URLs to storm products}

\item{product}{specific product to parse}
}
\description{
Get and parse product contents for each links
}
\keyword{internal}
