% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storms.R
\name{extract_storms}
\alias{extract_storms}
\title{extract_storms}
\usage{
extract_storms(basin, link)
}
\arguments{
\item{basin}{AL or EP}

\item{link}{URL to basin archive page}
}
\value{
4xN Dataframe
}
\description{
Extract storms for the given basin
}
\keyword{internal}
