% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{extract_year_archive_link}
\alias{extract_year_archive_link}
\title{extract_year_archive_link}
\usage{
extract_year_archive_link(link)
}
\arguments{
\item{link}{URL of archive page}
}
\value{
year 4-digit numeric
}
\description{
Extracts the year from the archive link.
}
\keyword{internal}
