% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_discus}
\alias{filter_discus}
\title{filter_discus}
\usage{
filter_discus(links)
}
\arguments{
\item{links}{Vector of URLs retrieved from storm's archive page.}
}
\description{
Filter out storm discussion links.
}
\keyword{internal}
