% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_forecast_advisories}
\alias{filter_forecast_advisories}
\title{filter_forecast_advisories}
\usage{
filter_forecast_advisories(links)
}
\arguments{
\item{links}{Vector of URLs retrieved from storm's archive page.}
}
\description{
Filter out forecast/advisory links
}

