% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_fstadv}
\alias{filter_fstadv}
\title{filter_fstadv}
\usage{
filter_fstadv(links)
}
\arguments{
\item{links}{Vector of URLs retrieved from storm's archive page.}
}
\description{
Filter out forecast/advisory links
}
\keyword{internal}
