% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_posest}
\alias{filter_posest}
\title{filter_posest}
\usage{
filter_posest(links)
}
\arguments{
\item{links}{Vector of URLs retrieved from storm's archive page.}
}
\description{
Get position estimates
}
\keyword{internal}
