% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_position_estimate}
\alias{filter_position_estimate}
\title{filter_position_estimate}
\usage{
filter_position_estimate(links)
}
\arguments{
\item{links}{Vector of URLs retrieved from storm's archive page.}
}
\description{
Get position estimates
}
