% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_prblty}
\alias{filter_prblty}
\title{filter_prblty}
\usage{
filter_prblty(links)
}
\arguments{
\item{links}{Vector of URLs retrieved from storm's archive page.}
}
\description{
Get strike probabilities.
}
\details{
Strike probability products were terminated at the end of the 2005
season, replaced by wind probabilities.
}
\seealso{
\code{\link{filter_wndprb}}
}
\keyword{internal}
