% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_products}
\alias{filter_products}
\title{filter_products}
\usage{
filter_products(ptn, links)
}
\arguments{
\item{ptn}{Regex pattern to find}

\item{links}{Vector of links to filter}
}
\description{
Filter out links matching pattern ptn
}
\keyword{internal}
