% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_public}
\alias{filter_public}
\title{filter_public}
\usage{
filter_public(links)
}
\arguments{
\item{links}{Vector of URLs retrieved from storm's archive page.}
}
\description{
Get public advisories
}
\keyword{internal}
