% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_public_advisories}
\alias{filter_public_advisories}
\title{filter_public_advisories}
\usage{
filter_public_advisories(links)
}
\arguments{
\item{links}{Vector of URLs retrieved from storm's archive page.}
}
\description{
Get public advisories
}

