% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_updates}
\alias{filter_updates}
\title{filter_updates}
\usage{
filter_updates(links)
}
\arguments{
\item{links}{Vector of URLs retrieved from storm's archive page.}
}
\description{
Get updates
}
\keyword{internal}
