% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_wind_probabilities}
\alias{filter_wind_probabilities}
\title{filter_wind_probabilities}
\usage{
filter_wind_probabilities(links)
}
\description{
Get wind probabilities
}
\details{
Wind probability products replaced the strike probability products 
at the beginning of the 2006 season.
}
\seealso{
\code{\link{filter_strike_probabilities}}
}

