% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_wndprb}
\alias{filter_wndprb}
\title{filter_wndprb}
\usage{
filter_wndprb(links)
}
\arguments{
\item{links}{Vector of URLs retrieved from storm's archive page.}
}
\description{
Get wind probabilities
}
\details{
Wind probability products replaced the strike probability products
at the beginning of the 2006 season.
}
\seealso{
\code{\link{filter_prblty}}
}
\keyword{internal}
