% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_extract_forecasts}
\alias{fstadv_extract_forecasts}
\title{fstadv_extract_forecasts}
\usage{
fstadv_extract_forecasts(content)
}
\arguments{
\item{content}{text of FORECAST/ADVISORY product}
}
\value{
list of forecasts
}
\description{
Extract all forecast data from FORECAST/ADVISORY product#'
}
\keyword{internal}
