% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_eye}
\alias{fstadv_eye}
\title{fstadv_eye}
\usage{
fstadv_eye(contents)
}
\arguments{
\item{contents}{text contents of FORECAST/ADVISORY}
}
\value{
numeric
}
\description{
Get eye diameter, if available
}

