% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_forecast_wind_radius}
\alias{fstadv_forecast_wind_radius}
\title{fstadv_forecast_wind_radius}
\usage{
fstadv_forecast_wind_radius(df)
}
\description{
Separate wind field columns to columns by quadrant.
}
\keyword{internal}
