% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_fwd_dir}
\alias{fstadv_fwd_dir}
\title{fstadv_fwd_dir}
\usage{
fstadv_fwd_dir(contents)
}
\arguments{
\item{contents}{Contents of forecast/advisory product.}
}
\value{
integer or NA
}
\description{
Extract forward direction from forecast/advisory product
}
