% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_fwd_mvmt}
\alias{fstadv_fwd_mvmt}
\title{fstadv_fwd_mvmt}
\usage{
fstadv_fwd_mvmt(contents, what = NULL)
}
\arguments{
\item{contents}{text contents of FORECAST/ADVISORY}

\item{what}{is being retrieved
\itemize{
  \item fwd_dir integer azimuth direction of movement (0 - 360)
  \item fwd_speed integer speed of movement in kts
}}
}
\value{
numeric
}
\description{
Get forward movement direction and speed
}
\details{
If STATIONARY should return NA
}
