% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_fwd_speed}
\alias{fstadv_fwd_speed}
\title{fstadv_fwd_speed}
\usage{
fstadv_fwd_speed(contents)
}
\arguments{
\item{contents}{Contents of forecast/advisory product.}
}
\value{
integer or NA
}
\description{
Extract forward speed from forecast/advisory product
}

