% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_gusts}
\alias{fstadv_gusts}
\title{fstadv_gusts}
\usage{
fstadv_gusts(contents)
}
\arguments{
\item{contents}{Contents of forecast/advisory product.}
}
\value{
integer or NA
}
\description{
Extract wind gusts from a forecast/advisory product.
}
