% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_lat}
\alias{fstadv_lat}
\title{fstadv_lat}
\usage{
fstadv_lat(contents)
}
\arguments{
\item{contents}{Content of forecast/advisory product}
}
\value{
numeric, positive if in northern hemisphere, negative for southern.
}
\description{
Extract latitude from forecast/advisory product
}

