% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_fstadv.R
\name{fstadv_lat_lon}
\alias{fstadv_lat_lon}
\title{fstadv_lat_lon}
\usage{
fstadv_lat_lon(contents)
}
\arguments{
\item{contents}{text contents of FORECAST/ADVISORY}
}
\value{
numeric
}
\description{
Returns numeric for latitude or longitude; negative if in
  southern or eastern hemisphere
}
\details{
Helper function to take character latitude or longitude and,
depending on the value of hemisphere return a positive or negative numeric,
or NA if not found.
}
\keyword{internal}
