% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_lon}
\alias{fstadv_lon}
\title{fstadv_lon}
\usage{
fstadv_lon(contents)
}
\arguments{
\item{contents}{Content of forecast/advisory product}
}
\value{
numeric, positive if in eastern hemisphere, negative for western.
}
\description{
Extract longitude from forecast/advisory product
}
\keyword{internal}
