% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_fstadv.R
\name{fstadv_pos_accuracy}
\alias{fstadv_pos_accuracy}
\title{fstadv_pos_accuracy()}
\usage{
fstadv_pos_accuracy(contents)
}
\arguments{
\item{contents}{text contents of FORECAST/ADVISORY}
}
\value{
numeric
}
\description{
Get position accuracy
}
\keyword{internal}
