% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_fstadv.R, R/fstadv.R
\name{fstadv_prev_pos}
\alias{fstadv_prev_pos}
\title{fstadv_prev_pos}
\usage{
fstadv_prev_pos(contents, adv_date)

fstadv_prev_pos(contents, adv_date)
}
\description{
Get storm's previous position

Get storm's previous position
}
\keyword{internal}
