% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_seas}
\alias{fstadv_seas}
\title{fstadv_seas}
\usage{
fstadv_seas(content, wind)
}
\arguments{
\item{content}{text of product}

\item{wind}{Wind value of current forecast/advisory product.}
}
\value{
boolean
}
\description{
There is only one line of sea data, 12FT seas in each quadrant. So this
  should go easier than the wind fields
}
\keyword{internal}
