% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_split}
\alias{fstadv_split}
\title{fstadv_split}
\usage{
fstadv_split(df, remove = TRUE)
}
\arguments{
\item{df}{Name (character) of dataframe holding default forecast/advisory
data.}

\item{remove}{Remove original dataframe after cleanup. Default TRUE.}
}
\description{
Takes the fstadv dataframe and splits into four relational
    dataframes; one for base data, one for wind radius data for the current
    product, one dataframe for each forecast position and one dataframe for
    wind radius data for each forecast position (within 72 hours).
}
\details{
The default dataframe for the \code{\link{get_fstadv}} contains 125
    columns. This function will take that dataframe and split it out to four
    narrow dataframes.
}
