% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_wind_radius}
\alias{fstadv_wind_radius}
\title{fstadv_wind_radius}
\usage{
fstadv_wind_radius(content)
}
\arguments{
\item{contents}{text of product}
}
\value{
dataframe
}
\description{
Parse wind radius data from product, if exists. This is somewhat
tricky as the wind fields are 64KT, 50KT and 34KT and are listed in
descending order. So the first line will not always be 64KT, 50KT or even
34KT depending on strength of storm. What I do here is just extract the
entire blob and work through it. I'll continue to look for ways to improve
it.

Complimentary to fstadv_get_wind_radius
}
\keyword{internal}
