% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_wind_radius_regex}
\alias{fstadv_wind_radius_regex}
\title{fstadv_wind_radius_regex}
\usage{
fstadv_wind_radius_regex(content)
}
\description{
Extra current wind radius from Forecast/Advisory product.
}
\keyword{internal}
