% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_winds}
\alias{fstadv_winds}
\title{fstadv_winds}
\usage{
fstadv_winds(contents)
}
\arguments{
\item{contents}{text contents of FORECAST/ADVISORY product}
}
\value{
numeric
}
\description{
Extract current maximum sustained winds from contents
}
