% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discus.R
\name{get_discus}
\alias{get_discus}
\title{get_discus}
\usage{
get_discus(link)
}
\arguments{
\item{link}{URL to storm's archive page.}
}
\description{
Return dataframe of discussion data.
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
    etc.}
  \item{Name}{Name of storm}
  \item{Adv}{Advisory Number}
  \item{Date}{Date of advisory issuance}
  \item{Contents}{Text content of product}
}
}
\examples{
\dontrun{
# Return dataframe of storm discussions for Tropical Storm Alex (AL011998)
get_discus("http://www.nhc.noaa.gov/archive/1998/1998ALEXadv.html")
}
}
\seealso{
\code{\link{get_storms}}, \code{\link{public}}
}
