% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{get_fstadv}
\alias{get_fstadv}
\title{get_fstadv}
\usage{
get_fstadv(link, msg = FALSE)
}
\arguments{
\item{link}{URL to storm's archive page.}

\item{msg}{Show link currently being worked. Default, FALSE.}
}
\description{
Return dataframe of forecast/advisory data.
}
\seealso{
\code{\link{get_storms}}, \code{\link{public}}
}
