% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storm_list.R
\name{get_ftp_storm_data}
\alias{get_ftp_storm_data}
\title{get_ftp_storm_data}
\usage{
get_ftp_storm_data(
  stormid,
  products = c("discus", "fstadv", "posest", "public", "prblty", "update", "wndprb")
)
}
\arguments{
\item{stormid}{A six-character alphanumeric string formatted as AABBCCCC
where
\describe{
  \item{AA}{The basin of the storm; AL or EP}
  \item{BB}{Storm number for the year as decimal number
    (e.g., 01, 02, ..., 10, ...)}
  \item{CCCC}{Year with century)}
}}

\item{products}{Products to retrieve; discus, fstadv, posest, public,
prblty, update, and windprb.}
}
\description{
Retrieve text products from the National Hurricane Center's FTP
  server. Not all products may exist for certain storms.
}
\seealso{
\code{\link{get_storm_data}}
}
