% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nhc_ftp_link}
\alias{get_nhc_ftp_link}
\title{get_nhc_ftp_link}
\usage{
get_nhc_ftp_link(withTrailingSlash = TRUE)
}
\arguments{
\item{withTrailingSlash}{True, by default. False returns URL without
trailing slash.}
}
\description{
Return root of NHC FTP server
}
\keyword{internal}
