% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prblty.R
\name{get_prblty}
\alias{get_prblty}
\title{get_prblty}
\usage{
get_prblty(link)
}
\arguments{
\item{link}{URL to storm's archive page.}
}
\description{
Return dataframe of strike probability data.
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
    etc.}
  \item{Name}{Name of storm}
  \item{Adv}{Advisory Number}
  \item{Date}{Date of advisory issuance}
  \item{Contents}{Text content of product}
}
}
\seealso{
\code{\link{get_storms}}, \code{\link{prblty}}
}
