% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storm_data.R
\name{get_product}
\alias{get_product}
\title{get_product}
\usage{
get_product(links, products)
}
\arguments{
\item{links}{a vector of links to storm pages}

\item{products}{vector of products}
}
\description{
This function acts as a hub for the individual product extraction
  functions. Given the product and links, it will begin the scraping
  process and return a dataset for that product.
}
\keyword{internal}
