% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storm_data.R
\name{get_product_links}
\alias{get_product_links}
\title{get_product_links}
\usage{
get_product_links(links, product)
}
\arguments{
\item{link}{url to a storm page}
}
\value{
vector of links for specific storm and product
}
\description{
get_product_links
}
