% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storm_data.R
\name{get_products}
\alias{get_products}
\title{get_products}
\usage{
get_products(link)
}
\arguments{
\item{link}{URL to storm's archive page.}
}
\description{
Get list of all products from a storm's archive page
}
