% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storm_data.R
\name{get_storm_content}
\alias{get_storm_content}
\title{get_storm_content}
\usage{
get_storm_content(link)
}
\arguments{
\item{link}{of archive page}
}
\value{
page content of archive page
}
\description{
Extract content from a storm's archive page
}
\keyword{internal}
