% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_track.R
\name{get_storm_track}
\alias{get_storm_track}
\title{get_storm_track}
\usage{
get_storm_track(
  serials,
  source = c("ACTIVE", "last3years", "since1980", "ALL", "EP", "NA", "NI", "SA", "SI",
    "SP", "WP")
)
}
\arguments{
\item{serials}{vector of serial numbers for a storm}

\item{source}{Short name for source, allows use of smaller file.}
}
\value{
data frame of storm track
}
\description{
get_storm_track
}
