% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storms.R
\name{get_storms_link}
\alias{get_storms_link}
\title{get_storms_link}
\usage{
get_storms_link(links, basins, p)
}
\arguments{
\item{links}{URLs to a year's archive page}

\item{basins}{Basins as passed to get_storms}

\item{p}{dplyr::progress_estimated object}
}
\description{
Get all links to storms from the annual archive pages
}
\keyword{internal}
