% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{get_update}
\alias{get_update}
\title{get_update}
\usage{
get_update(link)
}
\arguments{
\item{link}{URL to storm's archive page.}
}
\description{
Return dataframe of cyclone update data.
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
    etc.}
  \item{Name}{Name of storm}
  \item{Date}{Date of advisory issuance}
  \item{Key}{Unique ID of cyclone}
  \item{Contents}{Text content of product}
}
}
\seealso{
\code{\link{get_storms}}, \code{\link{update}}
}
