% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gis_advisory}
\alias{gis_advisory}
\title{gis_advisory}
\usage{
gis_advisory(key, advisory = as.character())
}
\arguments{
\item{key}{Key of storm (i.e., AL012008, EP092015)}

\item{advisory}{Advisory number. If NULL, all advisories are returned.
Intermediate advisories are acceptable.}
}
\description{
Advisory Forecast Track, Cone of Uncertainty, and
Watches/Warnings
}
\seealso{
\code{\link{gis_download}}
}
