% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gis_breakpoints}
\alias{gis_breakpoints}
\title{gis_breakpoints}
\usage{
gis_breakpoints()
}
\description{
Return link to breakpoints shapefile by year
}
\details{
Coastal areas placed under tropical storm and hurricane watches and
  warnings are identified through the use of "breakpoints." A tropical
  cyclone breakpoint is defined as an agreed upon coastal location that can
  be chosen as one of two specific end points or designated places between
  which a tropical storm/hurricane watch/warning is in effect. The U.S.
  National Weather Service designates the locations along the U.S. East,
  Gulf, and California coasts, Puerto Rico, and Hawaii. These points are
  listed in NWS Directive 10-605 (PDF). Individual countries across the
  Caribbean, Central America, and South America provide coastal locations
  for their areas of responsibility to the U.S. National Weather Service for
  the National Hurricane Center's use in tropical cyclone advisories when
  watches/warnings are issued by international partners. The National
  Hurricane Center maintains a list of pre-arranged breakpoints for the U.S.
  Atlantic and Gulf coasts, Mexico, Cuba and the Bahamas. Other sites are
  unofficial and sites not on the list can be selected if conditions warrant.
}
