% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gis_download}
\alias{gis_download}
\title{gis_download}
\usage{
gis_download(key, advisory = 1, destdir = tempdir())
}
\arguments{
\item{key}{Key of storm (i.e., AL012008, EP092015)}

\item{advisory}{Advisory number. If NULL, all advisories are returned. Intermediate
advisories are acceptable.}

\item{destdir}{Directory to save shapefile data.}
}
\description{
Get GIS data for storm.
}
\details{
Only available for storms >= 2008.
}
