% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gis_latest}
\alias{gis_latest}
\title{gis_latest}
\usage{
gis_latest(basins = c("AL", "EP"), destdir = tempdir())
}
\arguments{
\item{basins}{AL and/or EP.}

\item{destdir}{Directory to save shapefile data. Default is tempdir()}
}
\description{
Latest GIS datasets for active cyclones
}
