% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gis_outlook}
\alias{gis_outlook}
\title{gis_outlook}
\usage{
gis_outlook(destdir = tempdir())
}
\arguments{
\item{destdir}{Directory to save shapefile data. Saved to tmp dir by default.}
}
\description{
Tropical Weather Outlook
}
