% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gis_prob_storm_surge}
\alias{gis_prob_storm_surge}
\title{gis_prob_storm_surge}
\usage{
gis_prob_storm_surge(key, products, datetime = NULL, nobs = ~length(x))
}
\arguments{
\item{key}{Key of storm (i.e., AL012008, EP092015)}

\item{products}{list of products and associated n values; psurge (0:20) or
esurge (10, 20, 30, 40, 50).}

\item{datetime}{Datetime in \%Y\%m\%d\%H format.}

\item{nobs}{specific observations to retrieve by formula or numeric vector.
By default, only the last observation for each product is returned.}
}
\description{
Probabilistic Storm Surge
}
\details{
Probabilistic Storm Surge Forecasts
}
\section{Products}{

\describe{
    \item{esurge}{The Tropical Cyclone Storm Surge Exceedences (P-Surge 2.5)
          data shows the probability, in percent, of a specified storm surge,
          including tides, exceeding the specified height, in feet, during
          the forecast period indicated. The 10 percent exceedence height,
          for example, is the storm surge height, including tides, above
          ground level (AGL) such that there is a 10 percent chance of
          exceeding it. The product is based upon an ensemble of Sea, Lake,
          and Overland Surge from Hurricanes (SLOSH) model runs using the
          National Hurricane Center (NHC) official advisory and accounts for
          track, size, and intensity errors based on historical errors and
          astronomical tide. Valid values are 10, 20, 30, 40 or 50.}
    \item{psurge}{The Tropical Cyclone Storm Surge Probabilities (P-Surge
          2.5) data shows the probability, in percent, of a specified storm
          surge occurring during the forecast period indicated. The product
          is based upon an ensemble of Sea, Lake, and Overland Surge from
          Hurricanes (SLOSH) model runs using the National Hurricane Center
          (NHC) official advisory and accounts for track, size, and intensity
          errors based on historical errors and astronomical tide. Valid
          values are 0:20.}
}
}

\section{Subsetting datasets}{

Unlike other GIS data, these datasets do not contain advisory numbers but
are grouped by date/time values in \%Y\%m\%d\%H format. Typically, valid
hours will be 00, 06, 12 and 18 UTC though this may not always be the case.
To help you get the dataset you need you can use datetime and nobs together.
There are two parameters you can pass to help narrow down the dataset you
want.
\describe{
    \item{datetime}{The datetime in character formatted as \%Y\%m\%d\%H; i.e.,
        "2016083000".  If you pass this value then the GIS dataset for August
        30, 2016, 00:00 UTC will be returned. If you leave off the hour then
        all data for Aug 30, 2016 will be returned; leave off the hour and
        day then all data for August will be returned. You cannot drop values
        on the left side; only from the right. This is to help ensure you get
        the dataset you are requesting.}
    \item{nobs}{A formula or numeric vector specifying the range of datasets
        to retrieve.}
}
See examples for more guidance on using datetime and nobs together.
}

\examples{
\dontrun{
# Return the last psurge0 product for storm AL092016
gis_storm_surge("AL092016", products = list("psurge" = 0))

# Return the first psurge0 and esurge10 product for storm AL092016
gis_storm_surge("AL092016", products = list("psurge" = 0, "esurge" = 10), nobs = 1)

# Return all psurge0 products for September 3, 2016, storm AL092016
gis_storm_surge("AL092016", products = list("psurge" = 0),
                datetime = "20160903", nobs = ~c(1:length(x)))

# Last five esurge40 and esurge50 products for month of September, storm AL092016
gis_storm_surge("AL092016", products = list("esurge" = c(40, 50)),
                datetime = "201609", nobs = ~c((length(x)-10):length(x)))
}
}
\seealso{
\href{http://www.nhc.noaa.gov/surge/psurge.php}{Tropical Cyclone Storm Surge Probabilities}
}
