% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gis_storm_surge_flood}
\alias{gis_storm_surge_flood}
\title{gis_storm_surge_flood}
\usage{
gis_storm_surge_flood(key, advisory = as.numeric(),
  products = c("inundation", "tidalmask"))
}
\arguments{
\item{key}{Key of storm (i.e., AL012008, EP092015)}

\item{advisory}{Advisory number. If NULL, all available advisories are
returned.}

\item{products}{indundation or tidalmask}
}
\description{
Potential Storm Surge Flooding (Inundation)
}
