% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{knots_to_mph}
\alias{knots_to_mph}
\title{knots_to_mph}
\usage{
knots_to_mph(x)
}
\arguments{
\item{x}{wind speed in knots}
}
\value{
x in miles per hour
}
\description{
convert knots (kt) to miles per hour (mph)
}
\examples{
knots_to_mph(65)
}
