% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{mb_to_in}
\alias{mb_to_in}
\title{mb_to_in}
\usage{
mb_to_in(x)
}
\arguments{
\item{x}{barometric pressure in mb}
}
\value{
x in inches
}
\description{
convert millibars (mb) to inches of mercury (in)
}
\examples{
mb_to_in(999)
}
