% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{month_str_to_num}
\alias{month_str_to_num}
\title{month_str_to_num}
\usage{
month_str_to_num(m)
}
\arguments{
\item{m}{Month abbreviated (SEP, OCT, etc.)}
}
\value{
numeric 1-12
}
\description{
Convert three-character month abbreviation to integer
}
