% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storm_data.R
\name{parse_product_contents}
\alias{parse_product_contents}
\title{concept for isolating this step}
\usage{
parse_product_contents(contents, products)
}
\description{
concept for isolating this step
}
\keyword{internal}
