% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wndprb.R
\name{parse_stations}
\alias{parse_stations}
\title{parse_stations}
\usage{
parse_stations(x)
}
\arguments{
\item{x}{URL of station list}
}
\description{
Parse probability station listings for each basin.
}
\details{
At the moment, documentation on the format is unavailable. The
format changed during the 2017/2018 offseason and now includes a
numeric first column and numeric fifth, sixth and seventh column. All
values are identical per column.

Additionally, as of publication, PATRICK AFB in the Atlantic data source
actually contains eight columns; this is noted in
\code{\link{al_prblty_stations}}. SALINA CRUZ in
\code{\link{ep_prblty_stations}} is short one column.

I see no issues with the extra or missing data as I am unsure the value of
the data to begin with. A warning will be given so the user is aware,
but the important pieces (Location, Lat, Lon) all seem good.
}
\keyword{internal}
