% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posest.R
\name{posest}
\alias{posest}
\title{posest}
\usage{
posest(contents)
}
\arguments{
\item{contents}{URL of a specific position estimate product}
}
\value{
Dataframe
}
\description{
Extrapolate data from Position Estimate products.
}
\details{
Given a direct link to a position estimate product, parse and return
data frame of values.
}
\seealso{
\code{\link{get_posest}}
}
\keyword{internal}
