% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{saffir}
\alias{saffir}
\title{saffir}
\usage{
saffir(x)
}
\description{
Return category of storm based on wind. Assumes storm is a
cyclone. Saffir-Simpson Hurricane Scale does not apply to non-tropical storms.
}
