% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_adv_num}
\alias{scrape_adv_num}
\title{scrape_adv_num}
\usage{
scrape_adv_num(header)
}
\arguments{
\item{header}{Header text of product.}
}
\description{
Scrape advisory number from header.
}
\seealso{
\code{\link{scrape_header}}
}
