% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_header}
\alias{scrape_header}
\title{scrape_header}
\usage{
scrape_header(contents, ptn_product_title, advisory_number = TRUE)
}
\arguments{
\item{contents}{text product contents}

\item{ptn_product_title}{Pattern of product title to match}

\item{advisory_number}{Default is true; set to false if product does not
have an advisory number.}
}
\description{
Extract status, name, and advisory from products header.
}
\keyword{internal}
