% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_key}
\alias{scrape_key}
\title{scrape_key}
\usage{
scrape_key(header)
}
\arguments{
\item{header}{Header text of product.}
}
\description{
Extract Key from header
}
\seealso{
\code{\link{scrape_header}}
}
