% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_name}
\alias{scrape_name}
\title{scrape_name}
\usage{
scrape_name(header)
}
\arguments{
\item{header}{Header text of product.}
}
\description{
Scrape name from product header.
}
\seealso{
\code{\link{scrape_header}}
}
\keyword{internal}
