% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_status}
\alias{scrape_status}
\title{scrape_status}
\usage{
scrape_status(header)
}
\arguments{
\item{header}{Header text of product.}
}
\description{
Scrape status from product header.
}
\seealso{
\code{\link{scrape_header}}
}

