% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_track.R
\name{serial_from_name}
\alias{serial_from_name}
\title{Get IDs for a named storm}
\usage{
serial_from_name(name)
}
\arguments{
\item{name}{Name of the storm}
}
\value{
A character vector of storm IDs.
}
\description{
Get IDs for a named storm
}
