% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss}
\alias{ss}
\title{GIS storm surge shapefile dataset for Hurricane Sandy (AL182012)}
\format{An object of class \code{list} of length 1.}
\source{
\url{http://www.nhc.noaa.gov/gis/archive_psurge_results.php?id=al18&year=2012&name=Hurricane\%20SANDY}
}
\usage{
ss
}
\description{
GIS storm surge shapefile dataset for Hurricane Sandy (AL182012)
}
\keyword{datasets}
