% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{tidy_fcst}
\alias{tidy_fcst}
\title{tidy_fcst}
\usage{
tidy_fcst(df)
}
\arguments{
\item{df}{fstadv dataframe object}
}
\description{
Tidy forecasts of a fstadv dataframe object.
}
\details{
Gathers all forecast points, converts to narrow dataframe and
removes rows that are NA. Use Key, Adv and Date to join with other tidy
dataframes.
}
