% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{tidy_fcst_wr}
\alias{tidy_fcst_wr}
\title{tidy_fcst_wr}
\usage{
tidy_fcst_wr(df)
}
\arguments{
\item{df}{fstadv dataframe object}
}
\description{
Tidy forecast wind radii of a fstadv dataframe object
}
\details{
Converts forecast wind radii fields to narrow, tidy dataframe. Use
Key, Adv and Date to join with other tidy dataframes.
}
