% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{tidy_wr}
\alias{tidy_wr}
\title{tidy_wr}
\usage{
tidy_wr(df)
}
\description{
Tidy current wind radius of a fstadv dataframe object.
}
\details{
Takes current wind radius variables in a fstadv dataframe object,
tidys the data and removes rows where all values are NA. Use Key, Adv and
Date to join with other tidy dataframes.
}
