% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{toproper}
\alias{toproper}
\title{toproper}
\usage{
toproper(s, strict = TRUE)
}
\arguments{
\item{s}{Word or phrase to translate.}

\item{strict}{TRUE by default. Will convert all upper case characters to
lower case. If FALSE, no conversion will be done.}
}
\description{
Convert a string to proper case.
}
\details{
Taken from \code{\link{chartr}} examples in base R. Will take a
  phrase or, in the case of this project, full name and status of a cyclone,
  e.g., "TROPICAL STORM ALEX" and return "Tropical Storm Alex".
}
\examples{
\dontrun{
    toproper("TROPICAL STORM ALEX")
    toproper("TROPICAL STORM ALEX", strict = FALSE)
}
}
\keyword{internal}
