% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{twoal}
\alias{twoal}
\title{twoal}
\usage{
twoal()
}
\description{
Atlantic Tropical Weather Outlook
}
\details{
This function parses the latest xml tropical weather outlook for
the Atlantic ocean. The core data is located in the `channel$item` element
where `title`, `description` and `pubDate` reside. `link` is also
available to point to the NHC website.
}
