% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{twoep}
\alias{twoep}
\title{twoep}
\usage{
twoep()
}
\value{
Returns nested list of xml nodes.
}
\description{
East Pacific Tropical Weather Outlook
}
\details{
This function parses the latest xml tropical weather outlook for
the east Pacific. The core data is located in the `channel$item` element
where `title`, `description` and `pubDate` reside. `link` is also
available to point to the NHC website.
}
\seealso{
\url{http://www.nhc.noaa.gov/xml/TWOEP.xml}
}
