% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{updates}
\alias{updates}
\title{updates}
\usage{
updates(link, msg = FALSE)
}
\arguments{
\item{link}{Link to a storm's specific updates advisory product.}

\item{msg}{Display each link as being worked; default is FALSE}
}
\value{
Dataframe
}
\description{
Parse cyclone update products
}
\details{
Given a direct link to a cyclone update product, parse and return
dataframe of values.
}
\seealso{
\code{\link{get_updates}}
}
