% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{validate_year}
\alias{validate_year}
\title{validate_year}
\usage{
validate_year(y)
}
\value{
numeric year(s)
}
\description{
Test if year is 4-digit numeric.
}
\examples{
validate_year(1990)
validate_year(1991:1995)
validate_year(c(1996, 1997, 1998))
validate_year(c('1999', '2000'))
validate_year(list(2001, '2002', '2003'))
\dontrun{
validate_year(199) # Generates error}
}
\keyword{internal}
