% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wndprb.R
\name{wndprb}
\alias{wndprb}
\title{wndprb}
\usage{
wndprb(contents)
}
\arguments{
\item{contents}{Link to a storm's specific wind probability product.}
}
\description{
Parse wind probability products
}
\details{
Given a direct link to a wind probability product, parse and return
dataframe of values.
}
\keyword{internal}
