% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wsp}
\alias{wsp}
\title{GIS wind speed probabilities for Hurricane Sandy (AL182012)}
\format{An object of class \code{list} of length 3.}
\source{
\url{http://www.nhc.noaa.gov/gis/archive_wsp.php}
}
\usage{
wsp
}
\description{
GIS wind speed probabilities for Hurricane Sandy (AL182012)
}
\keyword{datasets}
