% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storms.R
\name{year_archives_link}
\alias{year_archives_link}
\title{year_archives_link}
\usage{
year_archives_link(years)
}
\arguments{
\item{year}{4-digit numeric}
}
\description{
Returns link to a year's archive page
}
\keyword{internal}
