% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wndprb.R
\name{cp_prblty_stations}
\alias{cp_prblty_stations}
\title{cp_prblty_stations}
\usage{
cp_prblty_stations()
}
\description{
Retrieve list of probability stations based in the central
    Pacific from the NHC. To be used in tandem with `wndprb` products.
}
