% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{create_df_fstadv}
\alias{create_df_fstadv}
\title{create_df_fstadv}
\usage{
create_df_fstadv()
}
\value{
empty dataframe
}
\description{
Template for forecast/advisories dataframe
}
\seealso{
\code{\link{get_fstadv}}
}
\keyword{internal}
