% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crul.R
\name{crul_get_storms}
\alias{crul_get_storms}
\title{crul_get_storms}
\usage{
crul_get_storms(years = format(Sys.Date(), "\%Y"), basins = c("AL", "EP"))
}
\arguments{
\item{years}{numeric or vector, four digits (\%Y format)}

\item{basins}{One or both of c("AL", "EP")}
}
\description{
Returns storms and product link.
}
