% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wndprb.R
\name{ep_prblty_stations}
\alias{ep_prblty_stations}
\title{ep_prblty_stations}
\usage{
ep_prblty_stations()
}
\description{
Retrieve list of probability stations based in the eastern
Pacific from the NHC. To be used in tandem with `wndprb` products.
}
\details{
Originally it was believed this data source would be removed by the
National Hurricane Center but it appears to have been updated. Additional
columns have been added, one up front and three in the back. These columns
all contain the same values each and I am unable to find documentation
describing the values.

Regardless, the data is kept, just in case.
}
\section{Warnings}{


Calling \code{ep_prblty_stations} will generate a warning:

> "Expected 7 pieces. Missing pieces filled with `NA` in 1 rows [41]."

Station SALINA CRUZ actually has six columns.
}

