% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storm_data.R
\name{extract_storm_links}
\alias{extract_storm_links}
\title{extract_storm_links}
\usage{
extract_storm_links(links, products)
}
\arguments{
\item{links}{data frame with URLs to a storm's archive page}

\item{products}{Products to return}
}
\description{
Extract product links from a storm's archive page
}
\keyword{internal}
