% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_strike_probabilities}
\alias{filter_strike_probabilities}
\title{filter_strike_probabilities}
\usage{
filter_strike_probabilities(links)
}
\description{
Get strike probabilities.
}
\details{
Strike probability products were terminated at the end of the 2005 
season, replaced by wind probabilities.
}
\seealso{
\code{\link{filter_wind_probabilities}}
}

