% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv}
\alias{fstadv}
\title{fstadv}
\usage{
fstadv(link, p)
}
\arguments{
\item{link}{URL of a specific FORECAST/ADVISORY product}

\item{p}{dplyr::progress_estimate.}
}
\description{
Extrapolate data from FORECAST/ADVISORY products.
}
\details{
Given a direct link to a forecast/advisory product, parse and
return dataframe of values.
}
\keyword{internal}
