% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv}
\alias{fstadv}
\title{fstadv}
\usage{
fstadv(link, display_link = TRUE)
}
\arguments{
\item{link}{URL of a specific FORECAST/ADVISORY product}

\item{display_link}{Display each link as being worked; default is TRUE.}
}
\value{
Dataframe
}
\description{
Extrapolate data from FORECAST/ADVISORY products.
}
\details{
Given a direct link to a forecast/advisory product, parse and 
return dataframe of values.
}
\seealso{
\code{\link{get_fstadv}}
}

