% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv}
\alias{fstadv}
\title{fstadv}
\usage{
fstadv(contents)
}
\arguments{
\item{contents}{URL of a specific FORECAST/ADVISORY product}
}
\value{
Data frame of values
}
\description{
Extrapolate data from FORECAST/ADVISORY products.
}
\details{
Given a direct link to a forecast/advisory product, parse and
   return data frame of values.
}
