% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_forecasts}
\alias{fstadv_forecasts}
\title{fstadv_fcst}
\usage{
fstadv_forecasts(content, key, adv, adv_date)
}
\arguments{
\item{content}{text content of FORECAST/ADVISORY}

\item{key}{Storm ID}

\item{adv}{Advisory Number}

\item{adv_date}{Date value of forecast/advisory product.}
}
\value{
boolean
}
\description{
Retrieve forecast data from FORECAST/ADVISORY products. Loads
  into respective dataframes (df_forecasts, df_forecast_winds)
}
\keyword{internal}
