% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_pressure}
\alias{fstadv_pressure}
\title{fstadv_pressure}
\usage{
fstadv_pressure(contents)
}
\arguments{
\item{contents}{text contents of FORECAST/ADVISORY product}
}
\value{
numeric
}
\description{
Return current minimum central pressure of storm in
    millibars (mb)
}
\keyword{internal}
