% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_fstadv.R
\name{fstadv_seas}
\alias{fstadv_seas}
\title{fstadv_seas}
\usage{
fstadv_seas(content)
}
\arguments{
\item{content}{text of product}
}
\value{
boolean
}
\description{
There is only one line of sea data, 12FT seas in each quadrant.
So this should go easier than the wind fields
}
\keyword{internal}
