% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{fstadv_winds_gusts}
\alias{fstadv_winds_gusts}
\title{fstadv_winds_gusts}
\usage{
fstadv_winds_gusts(contents)
}
\arguments{
\item{contents}{text contents of FORECAST/ADVISORY product}
}
\value{
numeric
}
\description{
Get winds or gusts in knots (KT)
}
\keyword{internal}
