% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storm_data.R
\name{get_fstadv}
\alias{get_fstadv}
\title{get_fstadv}
\usage{
get_fstadv(link)
}
\arguments{
\item{link}{URL to storm's archive page.}
}
\description{
Return dataframe of forecast/advisory data.
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane, 
    etc.}
  \item{Name}{Name of storm}
  \item{Adv}{Advisory Number}
  \item{Date}{Date of advisory issuance}
  \item{Contents}{Text content of product}
}
}
\seealso{
\code{\link{get_storms}}, \code{\link{public}}
}

