% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_nhc_link}
\alias{get_nhc_link}
\title{get_nhc_link}
\usage{
get_nhc_link(withTrailingSlash = TRUE)
}
\arguments{
\item{withTrailingSlash}{True, by default. False returns URL without 
trailing slash.}
}
\description{
Return root link of NHC archive pages.
}

