% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posest.R
\name{get_posest}
\alias{get_posest}
\title{get_posest}
\usage{
get_posest(link)
}
\arguments{
\item{link}{URL to storm's archive page.}
}
\description{
Return dataframe of position estimate data.
}
\details{
This product was discontinued after the 2013 hurricane season and is
now included in the Tropical Cyclone Update product (\code{\link{update}}).
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
    etc.}
  \item{Name}{Name of storm}
  \item{Date}{Date of advisory issuance}
  \item{Contents}{Text content of product}
}
}
\seealso{
\code{\link{get_storms}}, \code{\link{posest}}
}
