% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prblty.R
\name{get_prblty}
\alias{get_prblty}
\title{get_prblty}
\usage{
get_prblty(link)
}
\arguments{
\item{link}{URL to storm's archive page.}
}
\description{
Strike probabilities; the chances of the center of a cyclone
passing within 65 nautical miles of a location.
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
    etc.}
  \item{Name}{Name of storm}
  \item{Adv}{Advisory Number}
  \item{Date}{Date of advisory issuance}
  \item{Location}{Location for which the probability statistics rely}
  \item{A}{Probability of a strike within the next 12 hours}
  \item{B}{Probability of a strike between 12 and 24 hours}
  \item{C}{Probability of a strike between 24 and 36 hours}
  \item{D}{Probability of a strike between 36 and 48 hours}
  \item{E}{Probability of a strike between 48 and 72 hours}
}
}
