% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{get_public}
\alias{get_public}
\title{get_public}
\usage{
get_public(link)
}
\arguments{
\item{link}{URL to storm's archive page.}
}
\description{
Return dataframe of public advisory data.
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
    etc.}
  \item{Name}{Name of storm}
  \item{Adv}{Advisory Number}
  \item{Date}{Date of advisory issuance}
  \item{Key}{Unique ID of the cyclone}
  \item{Contents}{Text content of product}
}
}
\seealso{
\code{\link{get_storms}}, \code{\link{public}}
}
