% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_track.R
\name{get_serial_numbers}
\alias{get_serial_numbers}
\title{Creates the serial numbers look up}
\usage{
get_serial_numbers()
}
\description{
This will create a fresh table for serial numbers
Since this is constantly updated it should be
refreshed regularly especially when seeking recent
tracks.
}
