% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storm_data.R
\name{get_storm_data}
\alias{get_storm_data}
\title{get_storm_data}
\usage{
get_storm_data(link, products = c("discus", "fstadv", "posest", "public",
  "prblty", "update", "wndprb"))
}
\arguments{
\item{link}{to storm's archive page.}

\item{products}{Products to retrieve; discus, fstadv, posest, public,
prblty, update, and windprb.}
}
\value{
list of dataframes for each of the products.
}
\description{
Retrieve data from products.
}
\details{
\code{get_storm_data} is a wrapper function to make it more
    convenient to access the various storm products.

Types of products:
\describe{
  \item{discus}{Storm Discussions. This is technical information on the
    cyclone such as satellite presentation, forecast model evaluation, etc.}
  \item{fstadv}{Forecast/Advisory. These products contain the meat of an
    advisory package. Current storm information is available as well as
    structural design and forecast data.}
  \item{posest}{Position Estimate. Issued generally when a storm is
    threatening; provides a brief update on location and winds.}
  \item{public}{Public Advisory. Issued for public knowledge; more often for
    Atlantic than East Pacific storms. Contains general information.}
  \item{prblty}{Strike Probability. Discontinued after the 2005 hurricane
    season, strike probabilities list the chances of x-force winds in a
    particular city.}
  \item{update}{Cyclone Update. Generally issued when a significant change
    occurs in the cyclone.}
  \item{windprb}{Wind Probability. Replace strike probabilities beginning in
    the 2006 season. Nearly identical.}
}

Progress bars are displayed by default. These can be turned off by setting
the dplyr.show_progress to FALSE. Additionally, you can display messages for
each advisory being worked by setting the rrricanes.working_msg to TRUE.
}
\examples{
\dontrun{
## Get public advisories for first storm of 2016 Atlantic season.
get_storms(year = 2016, basin = "AL") \%>\%
    slice(1) \%>\%
    .$Link \%>\%
    get_storm_data(products = "public")
## Get public advisories and storm discussions for first storm of 2017 Atlantic season.
get_storms(year = 2017, basin = "AL") \%>\%
    slice(1) \%>\%
    .$Link \%>\%
    get_storm_data(products = c("discus", "public"))
}
}
