% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storm_data.R
\name{get_storm_data}
\alias{get_storm_data}
\title{get_storm_data}
\usage{
get_storm_data(..., names = list(), link)
}
\arguments{
\item{...}{Products to retrieve. c("discus", "fstadv", "posest", "public",
"prblty", "update", "windprb")}

\item{names}{By default product dataframes will be returned as shown in
\code{...}. The names parameter gives a way to provide alternative names for
the returned dataframes. Pass a named list where the name of each element
is that of the product. See examples for more information.}

\item{link}{to storm's archive page.}
}
\value{
Dataframes for each of the products.
}
\description{
Retrieve data from products.
}
\details{
\code{get_storm_data} is a wrapper function to make it more
convenient to access the various storm products.

Types of products:
\describe{
  \item{discus}{Storm Discussions. This is technical information on the
    cyclone such as satellite presentation, forecast model evaluation, etc.}
  \item{fstadv}{Forecast/Advisory. These products contain the meat of an
    advisory package. Current storm information is available as well as
    structural design and forecast data.}
  \item{posest}{Position Estimate. Issued generally when a storm is
    threatening; provides a brief update on location and winds.}
  \item{public}{Public Advisory. Issued for public knowledge; more often for
    Atlantic than East Pacific storms. Contains general information.}
  \item{prblty}{Strike Probability. Discontinued after the 2005 hurricane
    season, strike probabilities list the chances of x-force winds in a
    particular city.}
  \item{update}{Cyclone Update. Generally issued when a significant change
    occurs in the cyclone.}
  \item{windprb}{Wind Probability. Replace strike probabilities beginning in
    the 2006 season. Nearly identical.}
}
}
\examples{
## Get public advisories for Tropical Storm Charley, 1998
\dontrun{
get_storm_data("public",
               link = "http://www.nhc.noaa.gov/archive/1998/1998CHARLEYadv.html")

## Same as above but give alternate name.
get_storm_data("public",
               names = list("public" = "al.1998.charley.public"),
               link = "http://www.nhc.noaa.gov/archive/1998/1998CHARLEYadv.html")
## Get forecast/advisory and storm discussion
get_storm_data("fstadv", "discus",
               names = list("fstadv" = "al.1998.charley.fstadv",
                            "discus" = "al.1998.charley.discus"),
               link = "http://www.nhc.noaa.gov/archive/1998/1998CHARLEYadv.html")
}
}
