% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{get_updates}
\alias{get_updates}
\title{get_updates}
\usage{
get_updates(link, msg = FALSE)
}
\arguments{
\item{link}{URL to storm's archive page.}

\item{msg}{Show link being worked. Default, FALSE.}
}
\description{
Return dataframe of cyclone update data.
\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane, 
    etc.}
  \item{Name}{Name of storm}
  \item{Adv}{Advisory Number}
  \item{Date}{Date of advisory issuance}
  \item{Contents}{Text content of product}
}
}
\seealso{
\code{\link{get_storms}}, \code{\link{updates}}
}
