% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{get_url_contents}
\alias{get_url_contents}
\title{get_url_contents}
\usage{
get_url_contents(links, p)
}
\arguments{
\item{p}{dplyr progress bar object}

\item{link}{URL to download}
}
\description{
Get contents from URL
}
\details{
This function primarily is reserved for extracting the contents of
the individual products \(thought it can be used in other instances\). Often,
there are timeout issues. This is an attempt to try to work around that.
}
\keyword{internal}
