% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wndprb.R
\name{get_wndprb}
\alias{get_wndprb}
\title{get_wndprb}
\source{
\url{http://www.nhc.noaa.gov/about/pdf/About_Windspeed_Probabilities.pdf}
}
\usage{
get_wndprb(link)
}
\arguments{
\item{link}{URL to storm's archive page.}
}
\description{
Return dataframe of wind speed probability data.
}
\details{
Wind Speed Probability product replaced Strike Probabilities product
    after the 2005 hurricane season. These products may not be issued for
    every advisory/cyclone.

\describe{
  \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
    etc.}
  \item{Name}{Name of storm}
  \item{Adv}{Advisory Number}
  \item{Date}{Date of advisory issuance}
  \item{Wind}{Minimum wind speed for which probabilities reference}
  \item{Wind12}{Probability of sustained `Wind` within 12 hours}
  \item{Wind24}{Probability of sustained `Wind` within 24 hours}
  \item{Wind24Cum}{Cumulative probability through 24 hours}
  \item{Wind36}{Probability of sustained `Wind` within 36 hours}
  \item{Wind36Cum}{Cumulative probability through 36 hours}
  \item{Wind48}{Probability of sustained `Wind` within 48 hours}
  \item{Wind48Cum}{Cumulative probability through 48 hours}
  \item{Wind72}{Probability of sustained `Wind` within 72 hours}
  \item{Wind72Cum}{Cumulative probability through 72 hours}
  \item{Wind96}{Probability of sustained `Wind` within 96 hours}
  \item{Wind96Cum}{Cumulative probability through 96 hours}
  \item{Wind120}{Probability of sustained `Wind` within 120 hours}
  \item{Wind120Cum}{Cumulative probability through 120 hours}
}
}
