% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gis_download}
\alias{gis_download}
\title{gis_download}
\usage{
gis_download(url, destdir = tempdir(), ...)
}
\arguments{
\item{url}{link to GIS dataset to download.}

\item{destdir}{The destination directory to download and extract the zip file}

\item{...}{additional parameters for file.path}
}
\description{
Get GIS data for storm.
}
