% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gis_download}
\alias{gis_download}
\title{gis_download}
\usage{
gis_download(url, destdir = tempdir())
}
\arguments{
\item{destdir}{Directory to save shapefile data.}

\item{URL}{of GIS dataset to download.}
}
\description{
Get GIS data for storm.
}
