% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gis_download}
\alias{gis_download}
\title{gis_download}
\usage{
gis_download(url, ...)
}
\arguments{
\item{url}{link to GIS dataset to download.}

\item{...}{additional parameters for simple features}
}
\description{
Get GIS data for storm.
}
