% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{gis_outlook}
\alias{gis_outlook}
\title{gis_outlook}
\usage{
gis_outlook()
}
\value{
URL to latest Tropical Weather Outlook GIS dataset.
}
\description{
Tropical Weather Outlook
}
\seealso{
\code{\link{gis_download}}
}
