% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_storm_data.R
\name{load_storm_data}
\alias{load_storm_data}
\title{load_storm_data}
\usage{
load_storm_data(years, basins = c("AL", "EP"), products = NULL)
}
\arguments{
\item{years}{Numeric vector of one or multiple years between 1998 and current
year.}

\item{basins}{One or both of AL and EP.}

\item{products}{NULL, one or many of discus, fstadv, posest, prblty, public, update
and wndprb. See \link{get_storm_data} for a description of each of the
products. If products is NULL then a summary table of the year's storms is
returned. If a product is requested but not returned then the product does
not exist. For example, prblty does not exist for storms after 2005. And
short-lived storms may not have wndprb products.}
}
\description{
Load storm and year data from data repository.
}
\details{
This function is designed to give quicker access to storm data and
may be modified in future releases.

Using any of the product functions scrapes data from the NHC archives.
However, this can be somewhat time-consuming. Using this function just takes
a different route to datasets that have already been scraped. So obtaining
data is significantly faster.

I will not guarantee current storms will be up-to-date. But I will work on a
setup that does so as quickly as possible.
}
\examples{
\dontrun{
# Load year summary data for 2017, both basins
load_storm_data(years = 2017)

# Load multiple years, forecast/advisory data for AL storms.
load_storm_data(years = 2015:2016, basins = "AL", products = "fstadv")

# Load fstadv and wndprb for 2015 Atlantic storms.
load_storm_data(years = 2015, basins = "AL", products = c("fstadv", "wndprb"))
}
}
