% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posest.R
\name{posest}
\alias{posest}
\title{posest}
\usage{
posest(link, p)
}
\arguments{
\item{link}{URL of a specific position estimate product}

\item{p}{dplyr::progress_estimate.}
}
\value{
Dataframe
}
\description{
Extrapolate data from Position Estimate products.
}
\details{
Given a direct link to a position estimate product, parse and return
dataframe of values.
}
\seealso{
\code{\link{get_posest}}
}
\keyword{internal}
