% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posest.R
\name{posest}
\alias{posest}
\title{posest}
\usage{
posest(l, display_link = TRUE)
}
\arguments{
\item{l}{URL of a specific position estimate product}

\item{display_link}{Display each link as being worked; default is TRUE.}
}
\value{
Dataframe
}
\description{
Extrapolate data from Position Estimate products.
}
\details{
Given a direct link to a position estimate product, parse and return 
dataframe of values.
}
\seealso{
\code{\link{get_posest}}
}

