% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{public}
\alias{public}
\title{public}
\usage{
public(l, display_link = TRUE)
}
\arguments{
\item{l}{Link to a storm's specific public advisory product.}

\item{display_link}{Display each link as being worked; default is TRUE.}
}
\value{
Dataframe
}
\description{
Parse Public Advisory products
}
\details{
Given a direct link to a public advisory product, parse and return 
dataframe of values.
}
\seealso{
\code{\link{get_public}}
}

