% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\docType{package}
\name{rrricanes}
\alias{rrricanes}
\alias{rrricanes-package}
\title{rrricanes}
\description{
rrricanes is a web-scraping library for R designed to deliver
hurricane data (past and current) into well-organized datasets. With these
datasets you can explore past hurricane tracks, forecasts and structure
elements.

Text products (Forecast/Advisory, Public Advisory, Discussions and
Probabilities) are only available from 1998 to current. An effort will be
made to add prior data as available.
}
\section{Getting Storms}{

List all storms that have developed by year and basin. Year must be in a
four-digit format (\%Y) and no earlier than 1998. Basin can be one or both
of Atlantic ("AL") or East Pacific ("EP").
\describe{
  \item{\code{\link{get_storms}}}{List all storms by year, basin}
}
}

\section{Getting Storm Data}{

To be written...
}

