% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\docType{package}
\name{rrricanes}
\alias{rrricanes}
\alias{rrricanes-package}
\title{rrricanes}
\description{
rrricanes is a web-scraping library for R designed to deliver
hurricane data (past and current) into well-organized datasets. With these
datasets you can explore past hurricane tracks, forecasts and structure
elements.

This documentation and additional help articles
\href{https://timtrice.github.io/rrricanes/}{can be found online}.

Text products (Forecast/Advisory, Public Advisory, Discussions and
Probabilities) are only available from 1998 to current. An effort will be
made to add prior data as available.
}
\section{Getting Storms}{

List all storms that have developed by year and basin. Year must be in a
four-digit format (\%Y) and no earlier than 1998. Basin can be one or both
of Atlantic ("AL") or East Pacific ("EP").
\describe{
  \item{\code{\link{get_storms}}}{List all storms by year, basin}
}
}

\section{Getting Storm Data}{


\code{\link{get_storm_data}} can be used to select multiple products,
multiple storms and from multiple basins. You can also use
\code{\link{get_fstadv}} for Forecast/Advisory products,
\code{\link{get_wndprb}} for Wind Speed Probabilities (cyclones >= 2006; for
cyclones <= 2005 use \code{\link{get_prblty}}).

Additional text products are:
\describe{
  \item{\code{\link{get_discus}}}{Storm Discussions}
  \item{\code{\link{get_posest}}}{Position Estimates. Rare and used generally
    for threatening cyclones.}
  \item{\code{\link{get_public}}}{Public Advisory. General non-structured
    information exists in these products.}
  \item{\code{\link{get_update}}}{Updates. Generally issued when a cyclone
    undergoes a sudden change that requires immediate notice.}
}

The products above may take some time to load if the NHC website is slow (as
is often the case, unfortunately). You can access post-scraped datasets with
\code{\link{load_storm_data}}. Currently only annual summaries,
forecast/advisory, strike probabilities and wind speed probability products
exist. As of this writing it is up-to-date but caution is advised for active
cyclones. Use the above functions for the most up-to-date data as a
fall-back.
}

\section{GIS Data}{


For enhanced plotting of storm data, several GIS datasets are available. The
core GIS functions return URLs to help you refine the data you wish to view.
(Some products will not exist for all storms/advisories). These products are:

\describe{
  \item{\code{\link{gis_advisory}}}{Past track, current position, forecast and wind radii}
  \item{\code{\link{gis_breakpoints}}}{Breakpoints for watches and warnings}
  \item{\code{\link{gis_latest}}}{All available GIS products for active cyclones}
  \item{\code{\link{gis_outlook}}}{Tropical Weather Outlook}
  \item{\code{\link{gis_prob_storm_surge}}}{Probabilistic Storm Surge}
  \item{\code{\link{gis_windfield}}}{Wind Radii}
  \item{\code{\link{gis_wsp}}}{Wind Speed Probabilities}
}

\code{\link{gis_download}} will download the datasets from the above
functions.

Some GIS datasets will need to be converted to dataframes to plot geoms. Use
\code{\link{shp_to_df}} to convert SpatialLinesDataFrames and
SpatialPolygonsDataFrames. SpatialPointsDataFrames can be converted using
\code{tibble::as_data_frame} targeting the @data object.
}

\section{Package Options}{


\code{dplyr.show_progress} displays the dplyr progress bar when scraping raw
product datasets. In \code{\link{get_storms}}, it is based on the number of
years being requested. In the product functions (i.e.,
\code{\link{get_fstadv}}) it is based on the number of advisories. It can be
misleading when calling \code{\link{get_storm_data}} because it shows the
progress of working through a storm's product advisories but will reset on
new products/storms.

\code{rrricanes.working_msg} is set to FALSE by default. When TRUE, it will
list the current storm, advisory and date being worked.

\code{rrricanes.http_timeout} will set a timeout value in seconds. Often
when scraping raw datasets the connection may time out. Use this option if
this becomes an issue.

\code{rrricanes.http_attempts} will control the maximum number of attempts to
get a dataset. Default is 3 but no more than 5 attempts are permitted. If
\code{rrricanes.http_timeout} is reached, `rrricanes` will reattempt until
the value of \code{rrricanes.http_attempts} is reached.

\code{rrricanes.http_sleep} controls how long to wait between multiple
attempts. Default is 3 seconds.
}

