% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{saffir}
\alias{saffir}
\title{saffir}
\usage{
saffir(x)
}
\arguments{
\item{x}{Vector of wind speed values.}
}
\description{
Return category of tropical cyclone based on wind. Saffir-
Simpson Hurricane Scale does not apply to non-tropical cyclones.
}
\examples{
saffir(c(32, 45, 70, 90, 110, 125, 140))
}
