% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_contents}
\alias{scrape_contents}
\title{scrape_contents}
\usage{
scrape_contents(link)
}
\arguments{
\item{link}{URL to product page}
}
\value{
Contents of product
}
\description{
Extract text product from HTML
}
\keyword{internal}
