% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_contents}
\alias{scrape_contents}
\title{scrape_contents}
\usage{
scrape_contents(link, display_link = TRUE)
}
\arguments{
\item{link}{URL to product page}

\item{display_link}{Show link currently being worked. TRUE by default.}
}
\value{
Contents of product
}
\description{
Extract text product from HTML
}

