% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_header}
\alias{scrape_header}
\title{scrape_name_header}
\usage{
scrape_header(contents, ret = NULL)
}
\arguments{
\item{contents}{text content of product}

\item{ret}{Choose either status, name, adv or date to return.
\describe{
  \item{status}{status of cyclone (Hurricane, Tropical Storm, Subtropical 
    Storm...)}
  \item{name}{return name of storm}
  \item{adv}{return advisory number}
  \item{date}{Date/Time of product issuance.}
}}
}
\value{
Returns integer advisory number, character status or name, or date 
  in \%F \%R (yyyy-mm-dd hh:mm) format, UTC time zone.
}
\description{
Extract values from products header.
}
\details{
A typical header will look like this:
\preformatted{
  ZCZC MIATCPAT1 ALL
  TTAA00 KNHC DDHHMM
  BULLETIN
  TROPICAL DEPRESSION ONE ADVISORY NUMBER   1
  NATIONAL WEATHER SERVICE MIAMI FL
  11 AM AST MON JUL 27 1998
}
The first three lines are irrelevant. This function is only concerned about 
returning Status, Name, Advisory Number, and Date/Time values.
}

