% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_track.R
\name{serial_from_basin_id}
\alias{serial_from_basin_id}
\title{serial_numbers <- get_serial_numbers()}
\usage{
serial_from_basin_id(basin_id)
}
\arguments{
\item{basin_id}{The basin id}
}
\description{
Get all serial numbers for a basin
}
