% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstadv.R
\name{tidy_fcst}
\alias{tidy_fcst}
\title{tidy_fcst}
\usage{
tidy_fcst(df)
}
\arguments{
\item{df}{fstadv dataframe object}
}
\description{
Tidy forecasts of a fstadv dataframe object.
}
\details{
Gathers all forecast points, tidies dataframe to make one row per
forecast position. Complete cases only. Use Key, Adv and Date to join with
other tidy dataframes.

\describe{
   \item{Key}{Unique identifier of cyclone}
   \item{Adv}{Advisory number}
   \item{Date}{Date and time of advisory}
   \item{FcstDate}{Forecast date and time in UTC}
   \item{Lat}{Forecast latitude}
   \item{Lon}{Forecast Longitude}
   \item{Wind}{Forecast wind in knots}
   \item{Gust}{Forecast gust in knots}
}
}
\examples{
\dontrun{
get_fstadv("http://www.nhc.noaa.gov/archive/1998/1998ALEXadv.html") \%>\%
    tidy_fcst()
}
}
