% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_chart.R
\name{tracking_chart}
\alias{tracking_chart}
\title{tracking_chart}
\usage{
tracking_chart(countries = TRUE, states = TRUE, res = 110, ...)
}
\arguments{
\item{countries}{Show country borders. Default TRUE.}

\item{states}{Show state boundaries. Default TRUE. Ignored if `countries` is
FALSE.}

\item{res}{Resolution of charts; 110 (1:110m), 50 (1:50m), 10 (1:10m).
Default is low. The higher the resolution, the longer the plot takes to
appear.}

\item{...}{Additional ggplot2::aes parameters}
}
\value{
Returns ggplot2 object that can be printed directly or have new
    layers added.
}
\description{
Build base tracking chart using ggplot
}
\seealso{
\code{\link[ggplot2]{aes}}
}
