% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{update}
\usage{
update(link, msg = FALSE)
}
\arguments{
\item{link}{Link to a storm's specific update advisory product.}

\item{msg}{Display each link as being worked; default is FALSE}
}
\value{
Dataframe
}
\description{
Parse cyclone update products
}
\details{
Given a direct link to a cyclone update product, parse and return
dataframe of values.
}
\seealso{
\code{\link{get_update}}
}
\keyword{internal}
