% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{update}
\usage{
update(contents)
}
\arguments{
\item{contents}{Link to a storm's specific update advisory product.}
}
\value{
Dataframe
}
\description{
Parse cyclone update products
}
\details{
Given a direct link to a cyclone update product, parse and return
dataframe of values.
}
\seealso{
\code{\link{get_update}}
}
\keyword{internal}
