% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wndprb.R
\name{wndprb}
\alias{wndprb}
\title{wndprb}
\usage{
wndprb(link, display_link = TRUE)
}
\arguments{
\item{link}{Link to a storm's specific wind probability product.}

\item{display_link}{Display each link as being worked; default is TRUE.}
}
\value{
Dataframe
}
\description{
Parse wind probability products
}
\details{
Given a direct link to a wind probability product, parse and return 
dataframe of values.
}
\seealso{
\code{\link{get_wndprb}}
}

