% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wndprb.R
\name{wndprb}
\alias{wndprb}
\title{wndprb}
\usage{
wndprb(link, msg = FALSE)
}
\arguments{
\item{link}{Link to a storm's specific wind probability product.}

\item{msg}{Display each link as being worked; default is FALSE}
}
\value{
Dataframe
}
\description{
Parse wind probability products
}
\details{
Given a direct link to a wind probability product, parse and return
dataframe of values.
}
\seealso{
\code{\link{get_wndprb}}
}
\keyword{internal}
